convert resolve to ts (tsx works inside child process)
